tagalong/requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
autopep8==1.5.7
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==35.0.0
Django==3.2.7
django-ckeditor==6.2.0
django-cleanup==5.2.0
django-cors-headers==3.11.0
django-crontab==0.7.1
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-phonenumber-field==5.2.0
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
h11==0.13.0
idna==3.3
outcome==1.1.0
phonenumberslite==8.12.33
Pillow==8.3.2
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.1
regex==2022.3.2
requests==2.27.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.8
webdriver-manager==3.5.2
wsproto==1.0.0
